chore(core): rearrange e2e utils
